[UPDATE] return notification list and split ws handlers (#14)
